requirements.txt
django
djangorestframework
psycopg2-binary
requests
python-dotenv
gspread
celery
redis
celery[redis]
